# docs/openapi/academic.yaml
openapi: 3.1.0
info: { title: SSE Academic API, version: 0.1.0 }
servers: [{ url: / }]
...
paths:
...
  /education-plans/{id}/requirements:
    get:
      tags: [Academic]
      operationId: listEducationPlanRequirements
      x-roles: [ADMIN, ACADEMIC_STAFF]
      x-object-permission: Doc dinh muc mon hoc cua phien ban GD3.
      x-idempotency: Read-only.
      x-owners: { backend: Academic, client: Mobile }
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Dinh muc mon hoc theo hoc ky, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/CurriculumRequirement' } } } } }
    put:
      tags: [Academic]
      operationId: saveEducationPlanRequirement
      x-roles: [ADMIN, ACADEMIC_STAFF]
      x-object-permission: Chi sua DRAFT hoac REVISION_REQUESTED; ban da gui duyet hoac cong bo la bat bien.
      x-idempotency: Upsert duy nhat theo planId, semesterId va subjectId.
      x-owners: { backend: Academic, client: Mobile }
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SaveCurriculumRequirementRequest' } } } }
      responses:
        '200': { description: Dinh muc sau khi luu, content: { application/json: { schema: { $ref: '#/components/schemas/CurriculumRequirement' } } } }
        '409': { $ref: '#/components/responses/ApiErrorResponse' }
...
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }
  responses:
    ApiErrorResponse: { description: Loi phan quyen, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  schemas:
...
    SaveCurriculumRequirementRequest:
      type: object
      required: [semesterId, gradeLevel, subjectId, weeklyPeriods]
      properties:
        semesterId: { type: string }
        gradeLevel: { type: string }
        subjectId: { type: string }
        weeklyPeriods: { type: integer, minimum: 1, maximum: 20 }
        totalPeriods: { type: [integer, 'null'], minimum: 1, maximum: 500 }
        startDate: { type: [string, 'null'], format: date }
        endDate: { type: [string, 'null'], format: date }
        examWindowStart: { type: [string, 'null'], format: date }
        examWindowEnd: { type: [string, 'null'], format: date }
        milestone: { type: [string, 'null'], maxLength: 1000 }
    CurriculumRequirement:
      type: object
      required: [id, planId, semesterId, gradeLevel, subjectId, subjectName, weeklyPeriods, totalPeriods]
      properties:
        id: { type: string }
        planId: { type: string }
        semesterId: { type: string }
        gradeLevel: { type: string }
        subjectId: { type: string }
        subjectName: { type: string }
        weeklyPeriods: { type: integer }
        totalPeriods: { type: integer }
        startDate: { type: [string, 'null'], format: date }
        endDate: { type: [string, 'null'], format: date }
        examWindowStart: { type: [string, 'null'], format: date }
        examWindowEnd: { type: [string, 'null'], format: date }
        milestone: { type: [string, 'null'] }
...
    ApiError:
      type: object
      required: [status, message]
      properties: { status: { type: integer }, message: { type: string }, path: { type: [string, 'null'] } }
security: [{ bearerAuth: [] }]